pyproject.toml
[tool.poetry]
name = "rzd-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "rzd_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
